Allow logical name to be set on additional buckets
AWS allows buckets to have a logical name as well as a physical name.
This feature is used by `upload_assets` to make it easier to perform
actions on a kind of bucket within an environment. This is also useful
for additional S3 buckets where we need a different bucket per
environment.

Backwards compatibility is maintained by introducing an additional yaml
key for the `logical-name` which overrides the resource name. If
it's not set the resource name reverts to the bucket name.